240205 Daily Quiz 수정

+++ src/dailyquiz240205/LargeAppliance.java
@@ -1,7 +1,9 @@
 package dailyquiz240205;
 
+import java.math.BigDecimal;
+
 public class LargeAppliance extends Product {
-    public LargeAppliance(String name, double price, double weight) {
+    public LargeAppliance(String name, BigDecimal price, double weight) {
         super(name, price, weight);
     }
